feat(mkt): pricing calculator
