+++ apps/mkt/app/(default)/page.tsx
@@ -3,9 +3,9 @@ import Cta from "@/components/cta";
 import FeaturesOverview from "@/components/features-overview";
 import FeaturesTraits from "@/components/features-traits";
 import FeaturesAutomations from "@/components/features-automations";
-import PricingTabs from "@/components/pricing-tabs";
 import ProblemStatements from "@/components/problem-statements";
 import FeaturesSlack from "@/components/features-slack-digests";
+import PricingSection from "@/components/pricing-section";
 
 export default function Home() {
   return (
@@ -16,7 +16,7 @@ export default function Home() {
       <FeaturesSlack />
       <FeaturesAutomations />
       <FeaturesTraits />
-      <PricingTabs />
+      <PricingSection />
       {/* <Testimonials /> */}
       <Cta />
     </>
