+++ apps/mkt/components/tooltip.tsx
@@ -40,7 +40,8 @@ export default function Tooltip({
       >
         <Transition
           show={open}
-          className="w-[12.5rem] text-xs bg-white text-dark-100 border border-zinc-200 px-3 py-2 rounded shadow-lg overflow-hidden mt-1"
+          as="div"
+          className="w-[12.5rem] text-xs bg-white text-dark-900 border border-zinc-200 px-3 py-2 rounded shadow-lg overflow-hidden mt-1"
           enter="transition ease-out duration-200 transform"
           enterFrom="opacity-0 translate-y-1"
           enterTo="opacity-100 translate-y-0"
